#593 select tags for browse view tests for secondary users
