+++ test/runner.js
@@ -6,7 +6,6 @@ require('./insights/browse.test.js');
 require('./insights/add-insight.test.js');
 require('./insights/add-tags-to-insight.test.js');
 require('./insights/edit-insight.test.js');
-require('./insights/add-tags-to-insight.test.js');
 require('./insights/browse.test.js');
 require('./insights/add-insight.test.js');
 
@@ -16,7 +15,6 @@ require('./challenges/edit-chal.test.js');
 require('./challenges/browse.test.js');
 require('./challenges/details-view.test.js');
 require('./challenges/add-tags-to-challenge.test.js');
-require('./challenges/select-tags-browse-view.test.js');
 
 // orgs
 require('./organisations/add-tags-to-org-view.test.js');
@@ -27,7 +25,6 @@ require('./organisations/browse-orgs.test.js');
 require('./organisations/org-details-view.test.js');
 require('./organisations/edit-org-view.test.js');
 require('./organisations/edit-org.test.js');
-require('./organisations/select-tags-browse-view.test.js');
 require('./organisations/add-logo.test.js');
 
 // people
@@ -40,7 +37,7 @@ require('./people/edit-settings.test.js');
 require('./people/add-user.test.js');
 
 // tags
-require('./browse/select-tags-browse-view.test.js');
+require('./browse/select-tags.test.js');
 
 require('./unit/getCancelUrl.js');
